Add files via upload
This notebook uses the Africapolis 2015 data to append a csv dataset containing longitudes and latitudes with a new attribute specifying whether each long/lat pair lies in an urban area or not. This task is completed by checking whether each point lies inside any of the polygons defining the borders of African cities, given by the Africapolis dataset: https://africapolis.org/data This dataset is automatically downloaded if it is not already present in the specified directory.